Refactor the SwiftUI View
